Fix http stats support for uwsgitop
If we use the --stats-http option in uWSGI, we need to query the stats
endpoint with an HTTP request. Simply opening a socket to it won't have
any effect and the read() call will return no data.

I've also added argument parsing via argparse since manually inspecting
sys.argv is a pain. argparse is in the stdlib since python2.7, so we're
not introducing any external dependency for python2.7+. I've also added
argparse to the list of install requirements to keep support for
python 2.6.